refactor: criando pacote api/dto/response/game/base e separando app/session em app/session/game e app/session/player
