--- backend/src/main/java/com/chess/api/dto/response/game/BoardDTO.java
@@ -1,5 +1,0 @@
-package com.chess.api.dto.response.game;
-
-public record BoardDTO(
-    String fen
-) {}
