+++ backend/src/main/java/com/chess/api/dto/response/game/base/BoardDTO.java
@@ -1,0 +1,5 @@
+package com.chess.api.dto.response.game.base;
+
+public record BoardDTO(
+    String fen
+) {}
