+++ backend/src/main/java/com/chess/api/dto/response/game/base/ClockDTO.java
@@ -1,4 +1,4 @@
-package com.chess.api.dto.response.game;
+package com.chess.api.dto.response.game.base;
 
 public record ClockDTO(
     long whiteTimeRemaining,
