+++ backend/src/main/java/com/chess/api/dto/response/game/base/MoveDTO.java
@@ -1,4 +1,4 @@
-package com.chess.api.dto.response.game;
+package com.chess.api.dto.response.game.base;
 
 public record MoveDTO(
     String san,
