+++ backend/src/main/java/com/chess/app/service/game/GamePersistenceService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.chess.app.service.user.UserService;
-import com.chess.app.session.GameSession;
+import com.chess.app.session.game.GameSession;
 import com.chess.domain.event.GameFinishedEvent;
 import com.chess.domain.exception.game.DuplicateGameException;
 import com.chess.domain.exception.game.GameNotFoundException;
